+++ ft_printf_c.c
@@ -6,13 +6,13 @@
 /*   By: tnam <tnam@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/28 16:15:28 by tnam              #+#    #+#             */
-/*   Updated: 2022/11/30 12:29:57 by tnam             ###   ########.fr       */
+/*   Updated: 2022/12/06 15:50:39 by tnam             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_printf_c(va_list vargs_ptr)
+int	ft_printf_c(va_list vargs_ptr)
 {
 	ft_putchar_fd((char)va_arg(vargs_ptr, int), 1);
 	return (1);
